Create module for example data funcs
